added bad request validations

+++ server/controllers/chat.js
@@ -3,6 +3,9 @@ const getChats = async (req, res) => {
     res.json(await chatService.getChats());
 };
 const getChat = async (req, res) => {
+    if(!req.params.id || !Number.isInteger(req.params.id)){
+        return res.status(400).json({ errors: ['Bad Request of Chat'] });
+    }
     const chat = await chatService.getChatById(req.params.id);
     if (!chat) {
         return res.status(404).json({ errors: ['Chat not found'] });
@@ -10,6 +13,9 @@ const getChat = async (req, res) => {
     res.json(chat);
 };
 const getChatMessages = async (req, res) => {
+    if(!req.params.id || !Number.isInteger(req.params.id)){
+        return res.status(400).json({ errors: ['Bad Request of Chat'] });
+    }
     const chatMessages = await chatService.getMessagesOfChat(req.params.id);
     if (!chatMessages) {
         return res.status(404).json({ errors: ['Chat Messages not found'] });
@@ -17,6 +23,9 @@ const getChatMessages = async (req, res) => {
     res.json(chatMessages);
 };
 const deleteChat = async (req, res) => {
+    if(!req.params.id || !Number.isInteger(req.params.id)){
+        return res.status(400).json({ errors: ['Bad Request of Chat'] });
+    }
     const chat = await chatService.deleteChatById(req.params.id);
     if (!chat) {
         return res.status(404).json({ errors: ['Chat not found'] });
@@ -25,6 +34,9 @@ const deleteChat = async (req, res) => {
 };
 const createChat = async (req, res) => {
     const me = "";
+    if(!req.params.sender || !me) {
+        return res.status(400).json({ errors: ['Bad Request user'] });
+    }
     const chat = await chatService.createChat(req.params.sender,me);
     if (!chat) {
         return res.status(404).json({ errors: ['error when creating chat'] });
